Update INSTALL.srt (line 186, package tempstorage)
It looks like there is an effort to remove it as a dependency, however as of v4.3 (current) tempstorage is needed to successfully build HTML docs (tested with a vanilla openbsd 6.7, python 3.7 (following steps above accept for renaming make cmds in the script to gmake - worked fine save for some warnings, but this package was missing in the venv the first time.  Ran find after bin/pip install tempstorage to install.